ADM-361 move buttonGroup(next button and back button) to MetricsStepper
