+++ frontend/__tests__/src/components/metrics/MetricsStepper/MetricsStepper.test.tsx
@@ -1,28 +1,33 @@
 import { fireEvent, render } from '@testing-library/react'
 import MetricsStepper from '@src/components/Metrics/MetricsStepper'
 import { Provider } from 'react-redux'
-import { NEXT, BACK, STEPS, EXPORT_BOARD_DATA, PROJECT_NAME_LABEL } from '../../../fixtures'
 import { setupStore } from '../../../utils/setupStoreUtil'
+import { BACK, EXPORT_BOARD_DATA, NEXT, PROJECT_NAME_LABEL, STEPS } from '../../../fixtures'
+
+const METRICS = 'Metrics'
+const EXPORT = 'Export'
+const stepperColor = 'rgba(0, 0, 0, 0.87)'
+let store = setupStore()
+
+beforeEach(() => {
+  store = setupStore()
+})
+
+const setup = () =>
+  render(
+    <Provider store={store}>
+      <MetricsStepper />
+    </Provider>
+  )
 
 describe('MetricsStepper', () => {
-  let store = setupStore()
-  beforeEach(() => {
-    store = setupStore()
-  })
-
-  const setup = () =>
-    render(
-      <Provider store={store}>
-        <MetricsStepper />
-      </Provider>
-    )
-
   it('should show metrics stepper', () => {
     const { getByText } = setup()
 
     STEPS.map((label) => {
       expect(getByText(label)).toBeInTheDocument()
     })
+
     expect(getByText(NEXT)).toBeInTheDocument()
     expect(getByText(BACK)).toBeInTheDocument()
   })
@@ -40,24 +45,25 @@ describe('MetricsStepper', () => {
 
     fireEvent.click(getByText(NEXT))
 
-    expect(getByText(PROJECT_NAME_LABEL)).toBeInTheDocument()
+    expect(getByText(METRICS)).toHaveStyle(`color:${stepperColor}`)
   })
 
-  it('should show metrics config step when click back button given metrics step', async () => {
+  it('should show metrics config step when click back button given metrics step', () => {
     const { getByText } = setup()
 
     fireEvent.click(getByText(NEXT))
     fireEvent.click(getByText(BACK))
+
     expect(getByText(PROJECT_NAME_LABEL)).toBeInTheDocument()
   })
 
-  it('should show metrics export step when click next button given export step', async () => {
+  it('should show metrics export step when click next button given export step', () => {
     const { getByText } = setup()
 
     fireEvent.click(getByText(NEXT))
     fireEvent.click(getByText(NEXT))
-
     fireEvent.click(getByText(EXPORT_BOARD_DATA))
-    expect(getByText(PROJECT_NAME_LABEL)).toBeInTheDocument()
+
+    expect(getByText(EXPORT)).toHaveStyle(`color:${stepperColor}`)
   })
 })
