+++ frontend/src/components/Metrics/ConfigStep/index.tsx
@@ -1,10 +1,9 @@
 import { FormControlLabel, Radio, RadioGroup } from '@mui/material'
 import React, { useState } from 'react'
-import { CHINA_CALENDAR, REGULAR_CALENDAR, STEPS } from '@src/constants'
+import { CHINA_CALENDAR, REGULAR_CALENDAR } from '@src/constants'
 import { DateRangePicker } from '@src/components/Metrics/ConfigStep/DateRangePicker'
-import { BackButton, ButtonGroup, ConfigStepWrapper, ExportButton, NextButton, ProjectNameInput } from './style'
+import { ConfigStepWrapper, ProjectNameInput } from './style'
 import { useAppDispatch, useAppSelector } from '@src/hooks/useAppDispatch'
-import { backStep, nextStep, selectStepNumber } from '@src/context/stepper/StepperSlice'
 import { MetricsTypeCheckbox } from '@src/components/Metrics/ConfigStep/MetricsTypeCheckbox'
 import { updateBoardVerifyState } from '@src/context/board/boardSlice'
 import {
@@ -16,19 +15,10 @@ import {
 
 export const ConfigStep = () => {
   const dispatch = useAppDispatch()
-  const activeStep = useAppSelector(selectStepNumber)
   const projectName = useAppSelector(selectProjectName)
   const calendarType = useAppSelector(selectCalendarType)
 
   const [isEmptyProjectName, setIsEmptyProjectName] = useState<boolean>(false)
-
-  const handleNext = () => {
-    dispatch(nextStep())
-  }
-
-  const handleBack = () => {
-    dispatch(backStep())
-  }
 
   return (
     <ConfigStepWrapper>
@@ -60,14 +50,6 @@ export const ConfigStep = () => {
       </RadioGroup>
       <DateRangePicker />
       <MetricsTypeCheckbox />
-      <ButtonGroup>
-        <BackButton onClick={handleBack}>Back</BackButton>
-        {activeStep === STEPS.length - 1 ? (
-          <ExportButton>Export board data</ExportButton>
-        ) : (
-          <NextButton onClick={handleNext}>Next</NextButton>
-        )}
-      </ButtonGroup>
     </ConfigStepWrapper>
   )
 }
