+++ frontend/src/components/Metrics/MetricsStepper/index.tsx
@@ -2,14 +2,24 @@ import React from 'react'
 import Box from '@mui/material/Box'
 import Stepper from '@mui/material/Stepper'
 import Step from '@mui/material/Step'
-import { MetricsStepperContent, MetricsStepLabel } from './style'
-import { useAppSelector } from '@src/hooks/useAppDispatch'
-import { selectStepNumber } from '@src/context/stepper/StepperSlice'
+import { MetricsStepperContent, MetricsStepLabel, ButtonGroup, NextButton, ExportButton, BackButton } from './style'
+import { useAppDispatch, useAppSelector } from '@src/hooks/useAppDispatch'
+import { backStep, nextStep, selectStepNumber } from '@src/context/stepper/StepperSlice'
 import { ConfigStep } from '@src/components/Metrics/ConfigStep'
 import { STEPS } from '@src/constants'
+import { MetricsStep } from '@src/components/Metrics/MetricsStep'
 
 const MetricsStepper = () => {
+  const dispatch = useAppDispatch()
   const activeStep = useAppSelector(selectStepNumber)
+
+  const handleNext = () => {
+    dispatch(nextStep())
+  }
+
+  const handleBack = () => {
+    dispatch(backStep())
+  }
 
   return (
     <Box>
@@ -23,8 +33,17 @@ const MetricsStepper = () => {
         })}
       </Stepper>
       <MetricsStepperContent>
-        <ConfigStep />
+        {activeStep === 0 && <ConfigStep />}
+        {activeStep === 1 && <MetricsStep />}
       </MetricsStepperContent>
+      <ButtonGroup>
+        <BackButton onClick={handleBack}>Back</BackButton>
+        {activeStep === STEPS.length - 1 ? (
+          <ExportButton>Export board data</ExportButton>
+        ) : (
+          <NextButton onClick={handleNext}>Next</NextButton>
+        )}
+      </ButtonGroup>
     </Box>
   )
 }
